Add a backcompat feature flag
Versions v0.7.1 and v0.7.2 relied exclusively on the .so files without .1 suffix.
This is 'bad' because by convention that signals we don't care about ABI compatibility.
However in weird cases (*ahum* Tauri bundled appimages) this .so file is the only one
available. Using this feature flag allows them some time to fix this problem and bundle
with the correct filename.